Move code actions handler
